refactor: configure links to serve static content through nginx on server
The browser is making requests assuming that the index is at the root
of the server. However, it is behind a proxy at a path off the root.
Thus, we configure nginx to serve static content by sending all requests
to a specific route. We will then set protocol for that route within the
nginx server block. Yay.